Switch to species-specific equivalence classes for gene IDs. (#3)
This treats both Ensembl and Entrez IDs with equal importance, allowing genes
to be included even if they are not present in both annotations. This improves
the chance of mapping user-supplied identifiers. We achieve this by defining
"equivalence classes" from all interlinked gene IDs, e.g., if ENS1 maps to
Entrez 1 and 2, and Entrez 1 maps to ENS1 and ENS2, then [ENS1, 1, 2, ENS2]
forms one class. Indirect links are accepted for simplicity's sake, otherwise
we'd have to define multiple versions of the same gene... ugh.

Identifier types for the equivalence classes are stored in separate files.
This means that clients only need to download files for identifier types that
they actually use. In addition, the gene information for each species is stored
in separate files. This motivates us to split up the prebuilt references by
species, otherwise the gene IDs would need extra information to determine which
species-specific gene list to index. Fortunately, this is all desirable, as it
ensures that clients only need to download stuff for species they want.